refactor: articles read actions
